% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{count_column_occurrences}
\alias{count_column_occurrences}
\title{Count the number of occurrences of each value in the specified column}
\usage{
count_column_occurrences(df, column)
}
\arguments{
\item{df}{A data.frame from which the column value occurrences are counted}

\item{column}{A character vector with the column's name that must be counted}
}
\value{
A data.frame with two columns: the column value (with same name)
and n, the number of occurrences
}
\description{
Count the number of occurrences of each value in the specified column
}
