% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{count_starwars_sexes}
\alias{count_starwars_sexes}
\title{Count the number of occurrences of each value in the "sex" column of
the dplyr::starwars dataset}
\usage{
count_starwars_sexes()
}
\value{
A data.frame with two columns: the column value (with same name)
and n, the number of occurrences
}
\description{
Count the number of occurrences of each value in the "sex" column of
the dplyr::starwars dataset
}
\examples{
count_starwars_sexes()
}
