% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources-production.R
\name{expanded_production_sources}
\alias{expanded_production_sources}
\title{Create a new dataframe where each row has a year range into one where each
row is a single year, effectively 'expanding' the whole year range}
\usage{
expanded_production_sources(production_sources)
}
\arguments{
\item{production_sources}{A tibble dataframe
where each row contains the year range}
}
\value{
A tibble dataframe where each row
corresponds to a single year for a given source
}
\description{
Create a new dataframe where each row has a year range into one where each
row is a single year, effectively 'expanding' the whole year range
}
\examples{
production_sources <- tibble::tibble(
  Name = c("a", "b", "c"),
  Info_Format = c("year", "partial_series", "year"),
  Timeline_Start = c(1, 1, 2),
  Timeline_End = c(3, 4, 5),
  Timeline_Freq = c(1, 1, 2),
  SACO_link = NA,
)
expanded_production_sources(production_sources)
}
