% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal-hello-world.R
\name{my_hello_world}
\alias{my_hello_world}
\title{Creates a "Hello world!" ASCII art with an animal}
\usage{
my_hello_world(said_by)
}
\arguments{
\item{said_by}{The animal that says "Hello world!"}
}
\value{
The string generated with the ASCII art
}
\description{
Creates a "Hello world!" ASCII art with an animal
}
\examples{
cat(my_hello_world("cow"))
cat(my_hello_world("whale"))
cat(my_hello_world("stegosaurus"))
}
